+++ coffee_maker_navigator_2/lib/di/di.dart
@@ -1,6 +1,4 @@
 export 'src/framework/dependency_container.dart';
-export 'src/framework/manager/dependency_container_manager.dart';
-export 'src/framework/manager/dependency_container_resolver.dart';
-export 'src/framework/manager/dependency_container_registrar.dart';
-export 'src/framework/dependency_container_subscriber.dart';
+export 'src/framework/dependency_container_manager.dart';
+export 'src/framework/dependency_container_subscription_mixin.dart';
 export 'src/framework/dependency_container_injector.dart';
